chore(fe): import variables from utils
